+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - feat: add json schema validation for api
 - feat: add json schema validation for testcase v2
+- feat: add json schema validation for testsuite
 
 **Changed**