+++ docs/CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - feat: add json schema validation for api
 - feat: add json schema validation for testcase v2
-- feat: add json schema validation for testsuite
+- feat: add json schema validation for testsuite v2
 
 **Changed**