+++ docs/CHANGELOG.md
@@ -11,6 +11,8 @@
 **Changed**
 
 - refactor: use loader.load_cases to validate test files
+- refactor: use is_test_path to check if path is valid json/yaml file or a existed directory
+- refactor: use is_test_content to check if data_structure is apis/testcases/testsuites
 
 ## 2.4.9 (2019-12-29)